Improved documentation and adjusted website color scheme and style
For every page except the main page, I decided to stick with black text on a white background, instead of alternating between black-on-white and white-on-black (because that can get irritating to look at. I also made the document outline less intrusive for desktop users (though I still have the expandable popup thing for mobile due to space limits).